# Memory and speed improvements