--- a/backend/bls377/fft.go
+++ b/backend/bls377/fft.go
@@ -40,11 +40,11 @@ func FFT(a []fr.Element, w fr.Element) {
 }
 
 func asyncFFT(a []fr.Element, w fr.Element, wg *sync.WaitGroup, splits uint, deferDone bool) {
-	if deferDone {
-		defer wg.Done()
-	}
 	n := len(a)
 	if n == 1 {
+		if deferDone {
+			wg.Done()
+		}
 		return
 	}
 	m := n >> 1
@@ -56,6 +56,9 @@ func asyncFFT(a []fr.Element, w fr.Element, wg *sync.WaitGroup, splits uint, def
 
 	// if m == 1, then next iteration ends, no need to call 2 extra functions for that
 	if m == 1 {
+		if deferDone {
+			wg.Done()
+		}
 		return
 	}
 
@@ -86,6 +89,10 @@ func asyncFFT(a []fr.Element, w fr.Element, wg *sync.WaitGroup, splits uint, def
 		wg.Add(1)
 		go asyncFFT(a[m:n], w, wg, splits, true)
 		asyncFFT(a[0:m], w, wg, splits, false)
+	}
+
+	if deferDone {
+		wg.Done()
 	}
 }