--- a/backend/bls377/groth16/prove.go
+++ b/backend/bls377/groth16/prove.go
@@ -37,12 +37,10 @@ type Proof struct {
 }
 
 var (
-	root        fr.Element
 	minusTwoInv fr.Element
 )
 
 func init() {
-	root.SetString(backend_bls377.RootOfUnityStr)
 	minusTwoInv.SetUint64(2)
 	minusTwoInv.Neg(&minusTwoInv).
 		Inverse(&minusTwoInv)
@@ -53,7 +51,7 @@ func Prove(r1cs *backend_bls377.R1CS, pk *ProvingKey, solution map[string]interf
 	proof := &Proof{}
 
 	// fft domain (computeH)
-	fftDomain := backend_bls377.NewDomain(root, backend_bls377.MaxOrder, r1cs.NbConstraints)
+	fftDomain := backend_bls377.NewDomain(r1cs.NbConstraints)
 
 	// sample random r and s
 	var r, s, _r, _s fr.Element