Main changes of this PR: 

1. Serialization changed from `gob` to `cbor` (using github.com/fxamacker/cbor) 
```
benchmark              old ns/op      new ns/op     delta
BenchmarkCircuit-8     1466841876     277256539     -81.10%
benchmark              old allocs     new allocs     delta
BenchmarkCircuit-8     3000434        24             -100.00%
benchmark              old bytes     new bytes     delta
BenchmarkCircuit-8     280080600     164103908     -41.41%
```
Serialization is still work in progress, but at least, this switch enables processing very large circuits.

2. Groth16 setup now uses `gurvy` `BatchScalarMultiplication` apis: 
```
benchmark                  old ns/op      new ns/op      delta
BenchmarkSetup/setup-8     5007385019     3949330653     -21.13%
benchmark                  old allocs     new allocs     delta
BenchmarkSetup/setup-8     3569126        99             -100.00%
benchmark                  old bytes     new bytes     delta
BenchmarkSetup/setup-8     241889728     69919128      -71.09%
```
We still do not consider the setup production ready, but again, this enables `gnark` to scale. 

3. Removed a lot of pointers in `ConstraintSystem`, changed data structures, and hunted some useless heap allocations. Here is an example of impact on the frontend Compile workflow: 
```
benchmark              old ns/op      new ns/op      delta
BenchmarkCircuit-8     5395826965     1824662889     -66.18%
```

4. Switch to latest `gurvy` , which yields significant speed improvements on the MultiExponentiation, Pairing and field arithmetics. 

5. Code cleaning and cosmetics.